+++ src/index.ts
@@ -3,6 +3,7 @@ export * from './api/get-widget-info';
 export * from './api/register-widget-configuration-screen';
 export * from './api/register-widget-task-handler';
 export * from './api/request-widget-update';
+export * from './api/request-widget-update-by-id';
 export * from './api/types';
 export * from './config-plugin.type';
 export * from './widgets/FlexWidget';